requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
typer>=0.9.0
loguru>=0.7.2
PyYAML>=6.0.1
scipy>=1.10.0
